SO-2154 Support all reference set types in SNOMED CT request API

+++ snomed/com.b2international.snowowl.snomed.importer.rf2/src/com/b2international/snowowl/snomed/importer/rf2/UnsupportedRf2RefSetMember.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2015 B2i Healthcare Pte Ltd, http://b2i.sg
+ * Copyright 2011-2017 B2i Healthcare Pte Ltd, http://b2i.sg
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -93,12 +93,12 @@ public abstract class UnsupportedRf2RefSetMember extends FakeEObjectImpl impleme
 	}
 
 	@Override
-	public void setMapGroup(byte value) {
+	public void setMapGroup(int value) {
 		throw new UnsupportedOperationException();
 	}
 
 	@Override
-	public void setMapPriority(byte value) {
+	public void setMapPriority(int value) {
 		throw new UnsupportedOperationException();		
 	}
